fix: 🐛 Change Text
