+++ src/components/room/style.module.scss
@@ -1,4 +1,4 @@
-.planning-pocker :global {
+.room :global {
   @apply mx-auto bg-abc-light-grey;
   .container {
     @apply py-80;
