+++ src/core-ui/modal-create-room/style.module.scss
@@ -1,4 +1,4 @@
-.create-new-room :global {
+.modal-create-room :global {
   @apply mx-auto;
   .container {
     @apply py-80;
